+++ tests/php73_files/src/001_array_key_first.php
@@ -1,0 +1,4 @@
+<?php
+var_export(array_key_first([]));
+var_export(array_key_first('not an array'));
+var_export(array_key_last(2));